Исправлен метод GetPropertyID для IInstruction GPAllowedStates переведён на List с Dictionary Исправлен подсчёт прогресса

+++ Project/Rybocompleks.GUI/Rybocompleks.Dispatcher/Classes/GrowingPlanCommon.cs
@@ -31,9 +31,9 @@ namespace Rybocompleks.GrowingPlan
                 if (currentInstructionTime <= required && nextInstructionTime > required)
                 {
                     AllowedStatesList[i].Progress =
-                        ((required.Hour - currentInstructionTime.Hour) * 60 +
+                        (Double)((required.Hour - currentInstructionTime.Hour) * 60 +
                         (required.Minute - currentInstructionTime.Minute)) /
-                        ((nextInstructionTime.Hour - currentInstructionTime.Hour) * 60 +
+                        (Double)((nextInstructionTime.Hour - currentInstructionTime.Hour) * 60 +
                         (nextInstructionTime.Minute - currentInstructionTime.Minute));
 
                     return AllowedStatesList[i];
